Extract wasmi_factor out of run_dna
