+++ core/src/nucleus/ribosome/mod.rs
@@ -6,6 +6,7 @@ pub mod capabilities;
 pub mod memory;
 mod run_dna;
 pub mod runtime;
+pub mod wasmi_factory;
 
 pub use self::{run_dna::*, runtime::*};
